Minor naming and indentation fixes
