+++ src/components/Data/Leftnav/IndexBranch.vue
@@ -7,7 +7,7 @@
       <span v-html="highlight(indexName, filter)"></span> ({{collectionCount}})
     </router-link>
     <ul class="collections">
-      <li v-for="collectionName in orderedFilteredStoredCollections">
+      <li v-for="collectionName in orderedFilteredStoredCollections" :key="collectionName">
         <router-link class="tree-item truncate"
            :to="{name: 'DataDocumentsList', params: {index: indexName, collection: collectionName}}"
            :class="{ 'active': isCollectionActive(indexName, collectionName) }">
@@ -15,7 +15,7 @@
           <span v-html="highlight(collectionName, filter)"></span>
          </router-link>
       </li>
-      <li v-for="collectionName in orderedFilteredRealtimeCollections">
+      <li v-for="collectionName in orderedFilteredRealtimeCollections" :key="collectionName">
         <router-link class="tree-item"
            :to="{name: 'DataCollectionWatch', params: {index: indexName, collection: collectionName}}"
            :class="{ 'active': isCollectionActive(indexName, collectionName) }">
@@ -29,7 +29,7 @@
 </template>
 
 <script>
-import {REMOVE_REALTIME_COLLECTION} from '../../../vuex/modules/index/mutation-types'
+import { REMOVE_REALTIME_COLLECTION } from '../../../vuex/modules/index/mutation-types'
 
 export default {
   props: {
@@ -44,39 +44,43 @@ export default {
     routeName: String,
     collections: Object
   },
-  data: function () {
+  data: function() {
     return {
       open: false
     }
   },
   computed: {
-    collectionCount () {
+    collectionCount() {
       if (!this.collections) {
         return 0
       }
 
       return this.collections.realtime.length + this.collections.stored.length
     },
-    filteredStoredCollections () {
+    filteredStoredCollections() {
       if (this.collections) {
-        return this.collections.stored.filter(col => col.indexOf(this.filter) !== -1)
+        return this.collections.stored.filter(
+          col => col.indexOf(this.filter) !== -1
+        )
       }
       return []
     },
-    filteredRealtimeCollections () {
+    filteredRealtimeCollections() {
       if (this.collections) {
-        return this.collections.realtime.filter(col => col.indexOf(this.filter) !== -1)
+        return this.collections.realtime.filter(
+          col => col.indexOf(this.filter) !== -1
+        )
       }
       return []
     },
-    orderedFilteredStoredCollections () {
+    orderedFilteredStoredCollections() {
       return this.filteredStoredCollections
         .map(obj => {
           return this.highlight(obj, this.filter)
         })
         .sort()
     },
-    orderedFilteredRealtimeCollections () {
+    orderedFilteredRealtimeCollections() {
       return this.filteredRealtimeCollections
         .map(obj => {
           return this.highlight(obj, this.filter)
@@ -85,11 +89,11 @@ export default {
     }
   },
   methods: {
-    toggleBranch () {
+    toggleBranch() {
       // TODO This state should be one day persistent across page refreshes
       this.open = !this.open
     },
-    getRelativeLink (isRealtime) {
+    getRelativeLink(isRealtime) {
       switch (this.routeName) {
         case 'DataCollectionWatch':
           return this.routeName
@@ -99,23 +103,27 @@ export default {
           return 'DataDocumentsList'
       }
     },
-    testOpen () {
+    testOpen() {
       if (this.currentIndex === this.indexName) {
         this.open = true
       }
     },
-    isIndexActive (indexName) {
+    isIndexActive(indexName) {
       return this.currentIndex === indexName && !this.currentCollection
     },
-    isCollectionActive (indexName, collectionName) {
-      return this.currentIndex === indexName && this.currentCollection === collectionName
+    isCollectionActive(indexName, collectionName) {
+      return (
+        this.currentIndex === indexName &&
+        this.currentCollection === collectionName
+      )
     },
-    highlight (value, filter) {
+    highlight(value, filter) {
       if (value && value !== '' && filter && filter !== '') {
         let index = value.toLowerCase().indexOf(filter.toLowerCase())
 
         if (index >= 0) {
-          value = value.substring(0, index) +
+          value =
+            value.substring(0, index) +
             '<strong class="highlight">' +
             value.substring(index, index + filter.length) +
             '</strong>' +
@@ -125,112 +133,106 @@ export default {
 
       return value
     },
-    removeRealtimeCollection (indexName, collectionName) {
-      this.$store.dispatch(REMOVE_REALTIME_COLLECTION, {index: indexName, collection: collectionName})
-      if (this.$route.params.index === indexName && this.$route.params.collection === collectionName) {
-        this.$router.push({name: 'DataIndexSummary', params: {index: indexName}})
+    removeRealtimeCollection(indexName, collectionName) {
+      this.$store.dispatch(REMOVE_REALTIME_COLLECTION, {
+        index: indexName,
+        collection: collectionName
+      })
+      if (
+        this.$route.params.index === indexName &&
+        this.$route.params.collection === collectionName
+      ) {
+        this.$router.push({
+          name: 'DataIndexSummary',
+          params: { index: indexName }
+        })
       }
     }
   },
   watch: {
-    currentIndex () {
+    currentIndex() {
       this.testOpen()
     },
-    currentCollection () {
+    currentCollection() {
       this.testOpen()
     }
   },
-  mounted () {
+  mounted() {
     this.testOpen()
   }
 }
 </script>
 
 <style scoped lang="css">
-ul.collections
-{
+ul.collections {
   padding-left: 15px;
   overflow-y: hidden;
   max-height: 0;
 }
 
-li
-{
+li {
   position: relative;
 }
 
 a,
-i.tree-toggle
-{
+i.tree-toggle {
   line-height: 32px;
   height: 32px;
 }
 
-a.tree-item
-{
+a.tree-item {
   padding: 0 5px;
   margin: 0 15px;
 }
 
-a.tree-item.active
-{
+a.tree-item.active {
   font-weight: bold;
   color: #3498db;
 }
 
-i.fa
-{
+i.fa {
   margin-right: 5px;
   color: rgb(100, 100, 100);
 }
 
-i.fa:hover
-{
+i.fa:hover {
   margin-right: 5px;
   color: rgb(50, 50, 50);
 }
 
-i.tree-toggle
-{
+i.tree-toggle {
   position: absolute;
   cursor: pointer;
   padding: 0 10px 0 10px;
   margin: 0 0 0 -10px;
-  transition-duration: .2s;
+  transition-duration: 0.2s;
   transform-origin: 50% 50%;
 }
 
 /** open-closed state **/
 
-.open i.tree-toggle
-{
+.open i.tree-toggle {
   transform: rotate(90deg);
 }
 
-.open ul.collections
-{
+.open ul.collections {
   max-height: 2000px;
   transition: max-height 0.5s ease-out;
 }
 
 /* webkit adjacent element selector bugfix */
-@media screen and (-webkit-min-device-pixel-ratio:0)
-{
-    .treeview
-    {
-        -webkit-animation: webkit-adjacent-element-selector-bugfix infinite 1s;
-    }
+@media screen and (-webkit-min-device-pixel-ratio: 0) {
+  .treeview {
+    -webkit-animation: webkit-adjacent-element-selector-bugfix infinite 1s;
+  }
 
-    @-webkit-keyframes webkit-adjacent-element-selector-bugfix
-    {
-        from
-        {
-            padding: 0;
-        }
-        to
-        {
-            padding: 0;
-        }
+  @-webkit-keyframes webkit-adjacent-element-selector-bugfix {
+    from {
+      padding: 0;
     }
+    to {
+      padding: 0;
+    }
+  }
 }
 </style>
