+++ src/components/Security/Roles/Filters.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="search-filter">
+  <div class="RolesFilters">
     <div class="card-panel card-header">
       <div class="row filters margin-bottom-0">
         <basic-filter @filters-basic-search="broadcastFilterBasicSearch">
@@ -10,20 +10,20 @@
 </template>
 
 <script>
-  import BasicFilter from './Filters/BasicFilter'
+import BasicFilter from './Filters/BasicFilter'
 
-  export default {
-    name: 'Filters',
-    props: {
-      basicFilter: Object
-    },
-    components: {
-      BasicFilter
-    },
-    methods: {
-      broadcastFilterBasicSearch (filters, sorting) {
-        this.$emit('filters-basic-search', filters, sorting)
-      }
+export default {
+  name: 'RolesFilters',
+  props: {
+    basicFilter: Object
+  },
+  components: {
+    BasicFilter
+  },
+  methods: {
+    broadcastFilterBasicSearch(filters, sorting) {
+      this.$emit('filters-basic-search', filters, sorting)
     }
   }
+}
 </script>
